refactor: remove GetInvolved
